+++ packages/atomic/src/components/search/facets/atomic-segmented-facet/atomic-segmented-facet.tsx
@@ -6,8 +6,10 @@ import {
 } from '../../../../utils/initialization-utils';
 import {
   buildFacet,
+  buildFacetConditionsManager,
   buildSearchStatus,
   Facet,
+  FacetConditionsManager,
   FacetOptions,
   FacetSortCriterion,
   FacetState,
@@ -21,7 +23,7 @@ import {FacetValuesGroup} from '../../../common/facets/facet-values-group/facet-
 import {FacetSegmentedValue} from '../facet-segmented-value/facet-segmented-value';
 import {Hidden} from '../../../common/hidden';
 import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
-import {BaseFacet} from '../../../common/facets/facet-common';
+import {parseDependsOn} from '../../../common/facets/facet-common';
 
 /**
  * The `atomic-segmented-facet` displays a horizontal facet of the results for the current query.
@@ -37,9 +39,7 @@ import {BaseFacet} from '../../../common/facets/facet-common';
   styleUrl: 'atomic-segmented-facet.pcss',
   shadow: true,
 })
-export class AtomicSegmentedFacet
-  implements InitializableComponent, BaseFacet<Facet>
-{
+export class AtomicSegmentedFacet implements InitializableComponent {
   @InitializeBindings() public bindings!: Bindings;
   public searchStatus!: SearchStatus;
   @State()
@@ -85,11 +85,28 @@ export class AtomicSegmentedFacet
    */
   @Prop({reflect: true}) public sortCriteria: FacetSortCriterion = 'automatic';
 
-  // TODO
+  /**
+   * The required facets and values for this facet to be displayed.
+   * Examples:
+   * ```html
+   * <atomic-segmented-facet facet-id="abc" field="objecttype" ...></atomic-segmented-facet>
+   *
+   * <!-- To show the facet when any value is selected in the facet with id "abc": -->
+   * <atomic-segmented-facet
+   *   depends-on-abc
+   *   ...
+   * ></atomic-segmented-facet>
+   *
+   * <!-- To show the facet when value "doc" is selected in the facet with id "abc": -->
+   * <atomic-facet
+   *   depends-on-abc="doc"
+   *   ...
+   * ></atomic-segmented-facet>
+   * ```
+   */
   @MapProp() @Prop() public dependsOn: Record<string, string> = {};
 
-  // TODO
-  @Prop({reflect: true}) public withSearch = true;
+  private dependenciesManager!: FacetConditionsManager;
 
   public initialize() {
     this.searchStatus = buildSearchStatus(this.bindings.engine);
@@ -103,8 +120,20 @@ export class AtomicSegmentedFacet
       injectionDepth: this.injectionDepth,
       hasBreadcrumbs: false,
     };
+
     this.facet = buildFacet(this.bindings.engine, {options});
     this.facetId = this.facet.state.facetId;
+    this.dependenciesManager = buildFacetConditionsManager(
+      this.bindings.engine,
+      {
+        facetId: this.facetId!,
+        conditions: parseDependsOn(this.dependsOn),
+      }
+    );
+  }
+
+  disconnectedCallback() {
+    this.dependenciesManager.stopWatching();
   }
 
   private renderValuesContainer(children: VNode[]) {