+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,51 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.107.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.106.1...@coveo/atomic@1.107.0) (2022-09-06)
+
+
+### Features
+
+* **insight:** added field defined condition props to insight result template ([#2391](https://github.com/coveo/ui-kit/issues/2391)) ([f26fc49](https://github.com/coveo/ui-kit/commit/f26fc492d8cef5261f155022507af3d34649cc0f))
+
+
+
+
+
+## [1.106.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.106.0...@coveo/atomic@1.106.1) (2022-09-02)
+
+
+### Bug Fixes
+
+* **atomic:** cleanup unused prop for segmented facet and add support for depends on ([#2390](https://github.com/coveo/ui-kit/issues/2390)) ([9d07f4f](https://github.com/coveo/ui-kit/commit/9d07f4f529c2009d758bcde952189543309c379c))
+
+
+
+
+
+# [1.106.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.105.1...@coveo/atomic@1.106.0) (2022-09-02)
+
+
+### Features
+
+* **atomic:** set default insight panel results per page ([#2385](https://github.com/coveo/ui-kit/issues/2385)) ([41c84e0](https://github.com/coveo/ui-kit/commit/41c84e09fd6b7a93fa9da86b0140c198ff8b7d87))
+
+
+
+
+
+## [1.105.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.105.0...@coveo/atomic@1.105.1) (2022-08-29)
+
+
+### Bug Fixes
+
+* **atomic:** fix no results message when it contains special characters ([#2381](https://github.com/coveo/ui-kit/issues/2381)) ([19bf75f](https://github.com/coveo/ui-kit/commit/19bf75f6dca1cd0dfdb03287785ee61fc1150d90))
+* **atomic:** text-align left for recent queries content ([#2382](https://github.com/coveo/ui-kit/issues/2382)) ([465278c](https://github.com/coveo/ui-kit/commit/465278c4704ad6ca18bcf46168f6b9900cceb2eb))
+
+
+
+
+
 # [1.105.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.104.1...@coveo/atomic@1.105.0) (2022-08-26)